Fix Collector TE Ticking when Multiblock is Broken
o It actually still ticks, but doesn't do anything.

While a TE will not be created until the appropriate blockstate property is set, the TE is NOT removed if the property is unset.  But perhaps this is just as well, as this lets us keep the TE data unless the actual core is broken, saving the player's progress if they accidentally break the multiblock.