Add Fluid Plug
o There is a degree of overlap between the CollectorFinder code and the code presently associated with cluster component deregistration.  It is worth reviewing (again) to see if the whole lot of it can be simplified / moved inside Finder.
o Network bits are duplicated between the EnergyPlug and FluidPlug.  They can be moved back into the superclass once there is some confidence that there is not another reason to keep them distinct.