Port to 1.16.4 -- Pass 2
Less commenting.  Worldgen works again.  There are some places (jungle pool, for example) that could do with being updated to use the Feature configuration, instead of hard coding the values.  However, since there are no plans to change the configuration beyond the existing 'on/off' values, this is not especially urgent.